Move TimedTask and PriorityTask classes to their own files.
