+++ track/sprint2/src/main/java/tasks/TimedTask.java
@@ -1,0 +1,17 @@
+package tasks;
+
+public class TimedTask extends Task {
+    private String deadline;
+
+    public TimedTask(String description, String deadline) {
+        super(description);
+        this.deadline = deadline;
+    }
+
+
+    @Override
+    public String getStatus() {
+        return super.getStatus() + "(due by: " + deadline + ")";
+    }
+
+}
